Update patch set 1
Patch Set 1:

It looks like you're assuming that the compiler wasn't smart enough to drop local variables before, but will be with the addition of a few { }? That seems odd, have you actually measured a decrease in stack usage?

In general, if we're hitting stack limits then we would either need to reuse variables explicitly, or else heap allocate them.

Patch-set: 1