Unconditionally disable NEON on the buggy CPU.
If the device with the known-broken NEON unit is detected, disable NEON even if
__ARM_NEON__ is set. Before, the default OPENSSL_armcap_P value was getting in
the way.

This is a somewhat fragile and silly state of affairs since Chrome, as of M50,
tells the compiler it has free reign to emit NEON code, yet we're trying to
account for a phone with a broken NEON unit. But, so far, we only have evidence
of the hand-written NEON tickling the bug and not the compiler-generated stuff,
so avoid the unintentional regression.

This is not the fix for M50, as in the bug, but it will fix master and M51. M50
will instead want to revert https://codereview.chromium.org/1730823002.

BUG=chromium:606629

Change-Id: I394f97fea2f09891dd8fa30e0ec6fc6b1adfab7a