Update patch set 2
Patch Set 2:

> Perhaps there's a reason that the compiler can't reuse values with
 > the previous code. In any case, the stack error is stack, right? So
 > we don't need to worry about it silently regressing?

I think it's because if the compiler can't see the definition of newhope_poly_uniform and friends, it doesn't know we don't stash a pointer to the NEWHOPE_POLY instance somewhere global. If we let it go out of scope, then it's free to assume that any stashed pointer isn't touched after that point.

Patch-set: 2