Update patch set 1
Patch Set 1:

> Is this moot given the expected update to the RFC?

(Oh, are you on that thread too?) Yeah, it seems the update is going to correctly put a MUST NOT in there. Though they have a completely different extension number and everything. I expect we'll have both for a time.

I think we should have a test one way or another, but we can certainly go back and enforce it for this extension too. We can probably get away with it. At a glance, NSS's server code seems to get it wrong, but no one uses NSS as a server. Anyone using the OpenSSL custom extensions API to do it will be fine because those don't run on resumption anyway.

Patch-set: 1
Reviewer: Gerrit User 5415 <5415@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0