Update patch set 2
Restored

Restoring. After poking about the unidirectional case, there's enough bugs around it that, with ssl3_read_bytes as it is now, I don't think it's worth supporting. Once we're in a better place, it will be a matter of deleting one check in SSL_read, so we can always make it work if needed. But I doubt anyone actually cares about close_notify.

Patch-set: 2
Status: new