Update patch set 3
Patch Set 3:

The fact that this is going via the dymanic linker suggests that it's not being compiled with -fvisibility=hidden. However, I poked at this and it appears that even with visibility=hidden, OS X will still go via the linker at run-time for hidden symbols, and missing symbols are a run-time failure.

Patch-set: 3