Update patch set 3
Patch Set 3:

> The fact that this is going via the dymanic linker suggests that
 > it's not being compiled with -fvisibility=hidden. However, I poked
 > at this and it appears that even with visibility=hidden, OS X will
 > still go via the linker at run-time for hidden symbols, and missing
 > symbols are a run-time failure.

That makes sense. Random system includes aren't annotated with exports, so the only way for -fvisibility=hidden to work would be if visibility were purely a property of the symbol definition and not the symbol use. So OS X's linker must be seeing a missing symbol and just assuming it will be filled in by a random shared object later.

Patch-set: 3