Remove in-place TLS record assembly for now.
Decrypting is very easy to do in-place, but encrypting in-place is a hassle
(the rules actually were wrong due to record-splitting). Take it out for now in
preparation for tightening the aliasing rules. If we decide to do in-place
encrypt later, probably it'd be more useful to return header + in-place
ciphertext + trailer. (That, in turn, needs a scatter/gather thing on the AEAD
thanks to TLS 1.3's padding and record type construction.)

Change-Id: I0187d39c541e76ef933b7c2c193323164fd8a156