Update patch set 3
Patch Set 3:

> I think that this is probably too complex. What's wrong with a
 > function that converts an ASN.1 signature into a plain format? Or
 > one that encodes an ECDSA_SIG into the plain format?

I think it's silly to be converting between the ASN.1 signature and the plain format (because we convert to and from ASN.1 for no reason other than that's all that the EVP interface supports). For my Token Binding use case, I'm fine using the lower level ECDSA_do_sign and ECDSA_do_verify functions to work with an ECDSA_SIG.

There are some other places in chromium (webcrypto is one) that want a raw ECDSA format but use the EVP interface, which are currently converting between ASN.1 and a plain format. I think those would benefit from having this in the EVP interface.

If you don't want this in EVP, what are your thoughts on having conversion functions between ECDSA_SIG and the plain format in ecdsa.h (like the existing ASN.1 ECDSA_SIG_from_bytes/ECDSA_SIG_to_bytes)?

Patch-set: 3