Update patch set 3
Patch Set 3: -Code-Review

> I think it's silly to be converting between the ASN.1 signature and
 > the plain format (because we convert to and from ASN.1 for no
 > reason other than that's all that the EVP interface supports).

Although in the abstract I agree, I'd still take it because a) it's just 64 bytes and b) it's simplier. It would be better if the plain format was the default and there was a function to convert to ASN.1, but that's not a viable option.

 > There are some other places in chromium (webcrypto is one) that
 > want a raw ECDSA format but use the EVP interface, which are
 > currently converting between ASN.1 and a plain format. I think
 > those would benefit from having this in the EVP interface.
 > 
 > If you don't want this in EVP, what are your thoughts on having
 > conversion functions between ECDSA_SIG and the plain format in
 > ecdsa.h (like the existing ASN.1 ECDSA_SIG_from_bytes/ECDSA_SIG_to_bytes)?

I'd have no problem with that.

Patch-set: 3
Reviewer: Gerrit User 5005 <5005@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0