Update patch set 1
Patch Set 1:

David, in another review, you mentioned that you were uncertain about the constant-timedness of this reduction. I'm wondering if it is even possible for this code to execute? After all, it seems like all the values have to be in the range [0, m) in order for this function to make sense. Maybe I'm missing something though?

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0