Make dtls1_do_handshake_write less stateful.
Now dtls1_do_handshake_write takes in a serialized form of the full message and
writes it. It's a little weird to take serialize and deserialize the header a
bunch, but msg_callback requires that we keep the full one around in memory
anyway. Between that and the handshake hash definition, DTLS really wants
messages to mean the assembled header, redundancies and all, so we'll just put
together messages that way.

This also fixes a bug where ssl_do_msg_callback would get passed in garbage
where the header was supposd to be. The buffered messages get sampled before
writing the fragment rather than after.

Change-Id: I4e3b8ce4aab4c4ab4502d5428dfb8f3f729c6ef9