Update patch set 1
Patch Set 1:

> https://tools.ietf.org/html/rfc7748#section-5 says that "Scalars
 > are assumed to be randomly generated bytes." Crypto has a bad
 > history of working around issues and subtly diverging from specs so
 > I think the barrier here should be quite high. What's the
 > motivation for this change?

I'm passing the result of X25519_keypair to trevp's curve25519_sign, which uses the private key without any further transformation. The result of my naive combination was that no signatures would successfully verify.

To be sure, that latter function should be changed to properly decode its input, but having hit the bug I proposed this change as well in the spirit of "conservative in what you produce" etc.

I also felt okay suggesting this since it matches the steps for "Computing secret keys" in djb's Curve25519 description.

Patch-set: 1