Update patch set 4
Patch Set 4:

I learned some more about this.

In OpenSSL, they have this:
https://github.com/openssl/openssl/blob/bb83c8796b9e708a881f7c49a2af7921e73b2298/crypto/bn/bn_exp.c#L14

Thus, in general, `alloca` is usually defined in OpenSSL. But, BoringSSL doesn't have that, so `alloca` isn't defined in BoringSSL. Thus, all the parts of the code that are guarded by `#ifdef alloca` are dead code. I verified this with my code coverage reports in *ring*.

So, at the very least, the `#ifdef alloca` parts should be removed. Or, if it is deemed more important to keep this code exactly as it is in OpenSSL upstream, we should perhaps add a note to that effect somewhere in the documentation.

Patch-set: 4
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0