Change SignatureAndHashAlgorithm to SignatureScheme in Go.
TLS 1.3 defines a new SignatureScheme uint16 enum that is backwards compatible
on the wire with TLS1.2's SignatureAndHashAlgorithm. This change updates the go
testing code to use a single signatureAlgorithm enum (instead of 2 separate
signature and hash enums) in preparation for TLS 1.3. It also unifies all the
signing around this new scheme, effectively backporting the change to TLS 1.2.

For now, it does not distinguish signature algorithms between 1.2 and 1.3
(RSA-PSS instead of RSA-PKCS1, ECDSA must match curve types). When the C code
is ready make a similar change, the Go code will be updated to match.

[Originally written by nharper, tweaked significantly by davidben.]

Change-Id: If9a315c4670755089ac061e4ec254ef3457a00de