Shush gcc uninitialized variable warning with optimizations.
I am honestly at a loss how gcc could possibly think this is uninitialized.
It's even got the GetInt implementation available. TestRShift is nearly
identical but doesn't flag. It seems to require *two* accesses of n in the ||
chain and that the two accesses be separated by a call to some other function.

(I've been able to repro it on 32-bit release mode.)

Change-Id: I23de7934b309f33a7c59ac182a43bbd6ba70f7c3