Explicitly mark nearly every test at TLS 1.2.
In preparation for TLS 1.3 using its actual handshake, switch most tests to TLS
1.3 and add liberal TODOs for the tests which will need TLS 1.3 variants.

In doing so, move a few tests from basic tests into one of the groups. Also
rename BadECDSACurve to BadECDHECurve (it was never ECDSA) and add a test to
make sure FALLBACK_SCSV is sensitive to correctly sensitive to the maximum
version.

Change-Id: Ifca6cf8f7a48d6f069483c0aab192ae691b1dd8e